YOPO/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
opencv-python==4.11.0.86
scipy==1.10.1
scikit-build==0.18.1
ruamel-yaml==0.17.21
numpy==1.22.3
tensorboard==2.14.0
open3d==0.19.0
rich==14.0.0
empy==4.2
catkin_pkg
rospkg
netifaces